test: refactor test helpers by moving CursorSpy and ConnectionSpy to a dedicated module and update imports across test files
